Feat : 웹툰 조회 일반 완성 - 전체조회, 필터, 제공자

+++ backend/manamana/src/main/java/com/webtoon/manamana/webtoon/dto/response/AdditionDTO.java
@@ -26,7 +26,7 @@ public class AdditionDTO {
         return AdditionDTO.builder()
                 .view(webtoonAddition.getView())
                 .scoreCount(webtoonAddition.getScoreCount())
-                .scoreAverage(String.format("%.2f", webtoonAddition.getTotalScore()/webtoonAddition.getScoreCount()))
+                .scoreAverage(String.format("%.2f", (double) webtoonAddition.getTotalScore()/(double)webtoonAddition.getScoreCount()))
                 .build();
     }
 }
